ange order of registration to chef in instructions
as we need to add drives and mount them prior to running the chef
with cassandra (the cassandra recipe starts cassandra) automatically
and then it if we run it before we add the drive we need to
move the cassandra data drive, mount the external disk, move the data
back and then start cassandra.

potentially we all these tasks could have beed done pripor to adding the
node to the cluster it would be better.